+++ CHANGELOG.md
@@ -59,6 +59,7 @@ high state of flux, you're at risk of it changing without notice.
     - make `isSubset` data-last (@gcanti)
     - make `elem` data-last (@gcanti)
     - make `union` data-last (@gcanti)
+    - make `intersection` data-last (@gcanti)
   - `Record`
     - make `isSubrecord` data-last (@gcanti)
     - make `lookup` data-last (@gcanti)
@@ -67,6 +68,7 @@ high state of flux, you're at risk of it changing without notice.
     - make `subset` data-last (@gcanti)
     - make `elem` data-last (@gcanti)
     - make `union` data-last (@gcanti)
+    - make `intersection` data-last (@gcanti)
   - `Semigroup`
     - make `fold` data-last (@gcanti)